fcache: keep track of path lengths
Now that ks_path keeps track of path lengths, track the length within
fcache. This can speed up lookup in the cache, as path length is
compared first.

Flatten the ks_path fields into ks_cached_file, and track LRU with
4 bits instead of 8 in order to keep the struct at 32 bytes, allowing
2 of them to fit within a cache line.

+++ include/kestrel.h
@@ -3,12 +3,13 @@
 #define __KESTREL_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 void setup_seccomp(void);
 
 struct ks_path {
 	const char *path;
-	size_t len;
+	uint32_t len;
 };
 
 struct ks_file {
